feat: apply uniqueness email validation
