+++ src/cmd/main.go
@@ -6,7 +6,7 @@ import (
 	"github.com/gofiber/fiber/v2"
 	userUseCase "github.com/rodrigoRVSN/beeus-api/src/application/use_case/user"
 	"github.com/rodrigoRVSN/beeus-api/src/config"
-	"github.com/rodrigoRVSN/beeus-api/src/infra/repository"
+	"github.com/rodrigoRVSN/beeus-api/src/infra/repository/user"
 	routes "github.com/rodrigoRVSN/beeus-api/src/infra/router"
 )
 
@@ -16,7 +16,7 @@ func main() {
 	config.LoadEnv()
 	db := config.ConnectDb()
 
-	userRepository := repository.NewUserRepository(db)
+	userRepository := userRepository.NewUserRepository(db)
 	userUseCase := userUseCase.NewUserUseCase(userRepository)
 
 	routes.SetupRoutes(app, userUseCase)
