+++ src/infra/repository/user/find_all_users.go
@@ -1,0 +1,15 @@
+package userRepository
+
+import "github.com/rodrigoRVSN/beeus-api/src/domain/entity"
+
+func (r *UserRepository) FindAllUsers(user *entity.User) ([]entity.User, error) {
+	users := []entity.User{}
+
+	response := r.DB.Find(&users)
+
+	if response.Error != nil {
+		return nil, response.Error
+	}
+
+	return users, nil
+}
