+++ src/infra/repository/user/user_repository.go
@@ -1,0 +1,13 @@
+package userRepository
+
+import (
+	"gorm.io/gorm"
+)
+
+type UserRepository struct {
+	DB *gorm.DB
+}
+
+func NewUserRepository(db *gorm.DB) *UserRepository {
+	return &UserRepository{DB: db}
+}
